add new api key button in api key dropwdown
